authorization - delete posts, PostPolicy

+++ app/Http/Controllers/PostController.php
@@ -43,6 +43,11 @@ class PostController extends Controller
     }
 
     public function destroy(Post $post){
+
+//        if (!$post->ownedBy(auth()->user)){
+//            dd('no');
+//        }
+        $this->authorize('delete', $post);
         $post->delete();
 
         return back();
